feat: create globalExceptionHandler
