+++ src/main/java/gift/model/ProductName.java
@@ -1,9 +1,5 @@
 package gift.model;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
 public class ProductName {
     private final String name;
 
@@ -38,11 +34,5 @@ public class ProductName {
 
     private boolean isContainsKakao(String name){
         return name.contains("카카오");
-    }
-
-    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(IllegalArgumentException.class)
-    public String handleIllegalArgumentException(IllegalArgumentException e) {
-        return e.getMessage();
     }
 }
