+++ src/main/java/gift/service/GlobalExceptionHandler.java
@@ -1,0 +1,40 @@
+package gift.service;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import javax.naming.AuthenticationException;
+import java.util.NoSuchElementException;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ResponseStatus(value = HttpStatus.NOT_FOUND)
+    @ExceptionHandler(NoSuchElementException.class)
+    public String handleNoSuchElementException(NoSuchElementException e){
+        return e.getMessage();
+    }
+
+    //잘못된 이름 입력
+    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(IllegalArgumentException.class)
+    public String handleIllegalArgumentException(IllegalArgumentException e) {
+        return e.getMessage();
+    }
+
+    //로그인, 실패
+    @ResponseStatus(value = HttpStatus.FORBIDDEN)
+    @ExceptionHandler(IllegalAccessException.class)
+    public String handleIllegalAccessException(IllegalAccessException e) {
+        return e.getMessage();
+    }
+
+    //유효하지 않은 토큰
+    @ResponseStatus(value = HttpStatus.UNAUTHORIZED)
+    @ExceptionHandler(AuthenticationException.class)
+    public String handleAuthenticationException(AuthenticationException e) {
+        return e.getMessage();
+    }
+}
