--- src/main/java/gift/service/ProductExceptionHandler.java
@@ -1,17 +1,0 @@
-package gift.service;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.NoSuchElementException;
-
-@ControllerAdvice
-public class ProductExceptionHandler {
-    @ResponseStatus(value = HttpStatus.NOT_FOUND)
-    @ExceptionHandler(NoSuchElementException.class)
-    public String handleNoSuchElementException(NoSuchElementException e){
-        return e.getMessage();
-    }
-}
